Introduce static Data::u16, etc., methods, and use them.
The goal is to solve a problem Norbert Preining encountered on Alpha
machines. I believe the critical portion of the commit is the one in
otfname.cc, which replaces USHORT_AT() for extracting 16-bit characters
from a P_MICROSOFT/E_MS_UNICODE_BMP string with Data::u16().
(USHORT_AT() assumes the input data is 16-bit aligned. Data::u16() does
not. In most cases we explicitly align input strings so they are as
aligned as we need them to be. But for strings, particularly strings
generated by random font tools, who knows?) The rest of the commit
is prophylactic. Also introduce configury for
--enable-precondition-checking, which can test alignment.