Ligature handling: Don't infinite loop.
Bug reported by Bob Tennent and Marc Penninga. Thanks to sloppy
coding, we sometimes would *raise* a score in one phase of the
"Repeat these steps until you reach a stable set of scores" loop;
the loop will only terminate if we only *lower* ligature scores.
Clean that up (and as a benefit only consider "ffi" and "ffl" as
conventional ff-ligatures [not e.g. "fix"], and tweak the scores
for more sensible results).